# perf: switch to japanese-reranker-tiny-v2 for faster query processing