## Summary

This PR switches from japanese-reranker-xsmall-v2 to japanese-reranker-tiny-v2 to achieve significant performance improvements with minimal accuracy trade-offs.

## Motivation

Query speed is critical for user experience in RAG systems. The tiny-v2 model provides 50-65% faster query processing while maintaining good accuracy, making it a better fit for TinyRAG's mission of providing lightweight, CPU-efficient RAG.

## Changes

- Updated `Reranker` class to use `hotchpotch/japanese-reranker-tiny-v2`
- Updated README documentation with tiny-v2 model specifications

## Performance Improvements

### Query Speed
- **JQaRA**: 0.612 → 0.217 sec/query (**64.5% faster**)
- **JaCWIR**: 0.424 → 0.198 sec/query (**53.3% faster**)

### Model Efficiency
- **Size**: 29.4M parameters (20% smaller than xsmall-v2)
- **Layers**: Only 3 layers (70% reduction from 10 layers)
- **Memory**: Lower memory footprint due to smaller model

### Accuracy
- **JQaRA**: NDCG 0.7944 → 0.7558 (minor decrease)
- **JaCWIR**: MAP/Hits remain perfect at 0.8000

## Benchmark Results

The tiny-v2 model maintains strong performance on Japanese benchmarks:
- JaCWIR: 0.9287
- JSQuAD: 0.9608  
- MIRACL: 0.7201
- Average: 0.8138

## Testing

- [x] Benchmarks run successfully with new model
- [x] Query speed improvements verified
- [x] Accuracy trade-offs are acceptable
- [x] All existing tests pass

## Breaking Changes

None. This is a drop-in replacement that maintains the same API.

🤖 Generated with [Claude Code](https://claude.ai/code)